[CM-1261] Add option to disable dismiss UI.
